add client display layout
