+++ TenderWatch/app/src/main/res/layout/activity_ntender_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.tenderWatch.ClientDetail">
+    tools:context="com.tenderWatch.NTenderDetail">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
@@ -14,13 +14,14 @@
         <android.support.v7.widget.Toolbar
             android:id="@+id/toolbar"
             android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
+            android:layout_height="wrap_content"
+            android:minHeight="?attr/actionBarSize"
             android:background="?attr/colorPrimary"
             android:theme="@style/GalaxyZooThemeToolbarDarkOverflow"
             app:popupTheme="@style/Theme.AppCompat.NoActionBar"  />
 
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/content_client_detail" />
+    <include layout="@layout/content_ntender_detail" />
 
 </android.support.design.widget.CoordinatorLayout>
